pyproject.toml
[tool.poetry]
name = "hesiod"
version = "0.1.0"
description = ""
authors = [
    "Luca De Luigi <lucadeluigi91@gmail.com>",
    "Riccardo Albertazzi <piano.vergato@gmail.com>"
]

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.3.1"
npyscreen = "^4.10.5"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.790"
pytest = "^6.0"
flake8 = "^3.8.4"
isort = "^5.6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']
exclude = '''
(
  /(
      .git
    | .mypy_cache
    | .venv
    | build
    | dist
  )
)
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 9
